emp-zk: switch length params to int64_t (full sweep)
Track emp-tool's int64_t length-parameter convention (see emp-tool's
top-level AGENTS.md rule). Length-typed parameters across emp-zk's
public ZK exec API and supporting helpers used a mix of `int`,
`size_t`, and `uint32_t`. The same footguns apply here as in emp-tool:

- `int` silently overflows at 2^31 elements.
- `size_t` underflows silently in `len -= batch` decrement-to-zero
  loops; one such hazard in emp-zk-arith/edabit/edabits.h is fixed by
  this commit's signed conversion.

Build-blocker fixes (required to compile against the new emp-tool):
- emp-zk-bool/bool_io.h: send_data_internal / recv_data_internal
  overrides take `int64_t` to match emp-tool/io/io_channel.h's new
  virtuals; send_bool_raw / recv_bool_raw helpers follow.
- emp-svole/fp_utility.h: the templated `uni_hash_coeff_gen<T>` and
  `vector_inn_prdt_sum_red<T>` overload-resolution-tied with
  emp-tool's runtime versions; previously the int-typed third arg
  let the non-template (emp-tool) win, but with emp-tool widened the
  template won and instantiated for T=block, where `mult_mod(block,
  block)` does not exist. Widening the template's `int sz` to
  `int64_t sz` restores tied conversion ranks and the non-template
  wins again for T=block.

Consistency sweep (~32 files):

emp-svole/: preot.h, mpfss_reg.h, fp_mpfss_reg.h, fp_cope.h,
fp_base_svole.h, fp_policy.h, f2k_policy.h, svole.h, lpn_amp.h,
fp_lpn.h, spfss_{sender,recver}.h, fp_spfss_{sender,recver}.h,
fp_utility.h.

emp-zk-arith/: zk_fp_exec.h (virtual base) + zk_fp_exec_prover.h +
zk_fp_exec_verifier.h, ostriple.h, polynomial.h, conversion.h,
int_fp.h, edabit/doub_auth_helper.h, edabit/edabits.h.

emp-zk-bool/: zk_bool_base.h, zk_bool_prover.h, zk_bool_verifier.h,
polynomial.h, emp-zk-bool.h, bool_io.h, ram-zk/ostriple.h,
ram-zk/poly_prdt.h, ram-zk/zk_ram.h, ram-zk/zk_ram_ext.h,
ram-zk/zk_ram_ro.h.

Function parameters widened: `int len`, `int length`, `int sz`,
`int size`, count-typed `int n` → `int64_t`. Matching loop indices
and intermediate sums follow. Virtual andgate_correctness_check's
task_n param widened uint32_t → int64_t in zk_bool_base + derived
and ram-zk/ostriple hierarchies (the FpOSTriple non-virtual ThreadPool
helper variant keeps uint32_t — bounded by CHECK_SZ = 1M).

NOT widened (intentional): party/thread/log-domain ints (`int party`,
`int threads`, `int log_bin_sz`, spfss `depth, leave_n`), template
non-type params, FpOSTriple ThreadPool helper's uint32_t lambda
captures, anything in third_party/ or kyber/.

All 25 emp-zk tests pass (40s wall).

+++ emp-zk/emp-svole/f2k_policy.h
@@ -56,7 +56,7 @@ struct F2kPolicy {
       if (party == BOB) delta = ferret->Delta;
     }
 
-    void extend(AuthValue<F2kPolicy> *out, int num) {
+    void extend(AuthValue<F2kPolicy> *out, int64_t num) {
       // Pull num*128 OTs out of the long-lived ferret session via
       // rcot_*_next chunks. Caller must have an open ferret session
       // (the bool backend opens one ctor->dtor; standalone tests have
@@ -64,7 +64,7 @@ struct F2kPolicy {
       std::vector<block> ferret_buffer((std::size_t)num * 128);
       const int64_t chunk = ferret->chunk_ots();
       std::vector<block> chunk_buf(chunk);
-      int64_t needed = (int64_t)num * 128, got = 0;
+      int64_t needed = num * 128, got = 0;
       // ferret_party = 3-party; ferret is the OT-sender exactly when
       // this side is BOB.
       const bool sender = (party == BOB);
